[fix] Fix startup task not working
