+++ back-end/src/internal/tasks/__init__.py
@@ -1,3 +1,4 @@
 """This module contains background tasks to be run by FastAPI's BackgroundTasks."""
 from .clean_orphaned_media import delete_orphan_images
 from .clean_orphaned_services import delete_orphan_services
+from .init_db import init_db
