+++ back-end/src/main.py
@@ -7,6 +7,7 @@ from fastapi.middleware.cors import CORSMiddleware
 from .config.config import config
 from .internal.auth import check_is_admin, get_current_user
 from .routers import auth, buckets, datasets, engines, experiments, iam, models
+from .internal.tasks import init_db
 
 with open(
     Path(__file__).parent.parent.joinpath("README.md"), "r", encoding="utf-8"
@@ -44,7 +45,8 @@ tags_metadata = [
     },
 ]
 app = FastAPI(
-    title="Model Zoo", description=description, openapi_tags=tags_metadata
+    title="Model Zoo", description=description, openapi_tags=tags_metadata,
+    on_startup=[init_db]
 )
 app.add_middleware(
     CORSMiddleware,
@@ -79,7 +81,6 @@ app.include_router(
 app.include_router(datasets.router, dependencies=[Depends(get_current_user)])
 app.include_router(iam.router, dependencies=[Depends(check_is_admin)])
 app.include_router(engines.router, dependencies=[Depends(get_current_user)])
-
 
 @app.get("/")
 def root():
